Вносит исправления по филдбеку
