+++ js/photosArray.js
@@ -1,0 +1,40 @@
+import { getRandomInteger, getRandomArrayElement, createIdGenerator } from './utils.js';
+import {
+  PICTURE_COUNT,
+  AVATAR_COUNT,
+  LIKE_MIN_COUNT,
+  LIKE_MAX_COUNT,
+  COMMENT_COUNT,
+  COMMENT_LINES,
+  DESCRIPTION_ARRAY,
+  NAMES_ARRAY
+} from './data.js';
+
+const generateCommentId = createIdGenerator();
+
+const createComment = () => ({
+  id: generateCommentId(),
+  avatar: `img/avatar-${getRandomInteger(1, AVATAR_COUNT)}.svg`,
+  message: getRandomArrayElement(COMMENT_LINES),
+  name: getRandomArrayElement(NAMES_ARRAY),
+});
+
+const generatePhotoId = createIdGenerator();
+
+const createPicture = (index) => ({
+  id: generatePhotoId(),
+  url: `photos/${index}.jpg`,
+  description: getRandomArrayElement(DESCRIPTION_ARRAY),
+  likes: getRandomInteger(LIKE_MIN_COUNT, LIKE_MAX_COUNT),
+  comments: Array.from(
+    { length: getRandomInteger(0, COMMENT_COUNT) },
+    createComment,
+  ),
+});
+
+const getPictures = () => Array.from(
+  { length: PICTURE_COUNT },
+  (_, pictureIndex) => createPicture(pictureIndex + 1),
+);
+
+export { getPictures };
